% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check.packages}
\alias{check.packages}
\title{Check Packages}
\usage{
check.packages(pkg)
}
\arguments{
\item{pkg}{List of packages to install.}
}
\description{
Check to see if packages are installed and installs them if they are not. Verifies install using require.
}
