% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs_and_covariances.R
\name{coefCovH}
\alias{coefCovH}
\title{Coefficients and Covariance Matrix (Cross Section Data)}
\usage{
coefCovH(linmod)
}
\arguments{
\item{linmod}{The regression model.}
}
\description{
Provides coefficients of a regression model and the covariance matrix of the coefficients. The calculation of the covariance matrix corrects for possible presence of heteroskedasticity using the method developed by White. This method is appropriate for cross-section data.
}
\section{Details}{

Let the regression model be: linmod=lm(y~x1+x2)

The function reports the 3 coefficients from the regression in the first column and the 3x3 covariance matrix of the coefficients in the next three columns.
}

