% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_draw_functions.R
\name{coefDrawH}
\alias{coefDrawH}
\title{coefDrawH}
\usage{
coefDrawH(linmod)
}
\arguments{
\item{linmod}{Linear regression model.}
}
\value{
mydraw A random draw for the coefficients.
}
\description{
Given a regression (linmod), this function takes a random draw for the coefficients.
}
