% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_draw_functions.R
\name{coefDrawHAC}
\alias{coefDrawHAC}
\title{coefDrawHAC}
\usage{
coefDrawHAC(linmod)
}
\arguments{
\item{linmod}{Linear regression model.}
}
\value{
mydraw A random draw for the coefficients.
}
\description{
Tthis function takes a random draw for the coefficients of a time series regression.
}
