% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_hypothesis_functions.R
\name{coefTestH}
\alias{coefTestH}
\title{Linear Hypothesis Test (Cross Section Data)}
\usage{
coefTestH(model, conditions)
}
\arguments{
\item{model}{Regression model being tested.}

\item{conditions}{Hypothesis being tested.}
}
\description{
Tests a linear hypothesis about coefficients. The calculation corrects for possible presence of heteroskedasticity using the method developed by White. This method is appropriate for cross-section data.
}
\section{Details}{

Let the regression model be: linmod=lm(y~x1+x2+x3+x4)

To test the hypothesis that the coefficient of x1 equals the coefficient of x2 and that the coefficient of x3 equals twice the coefficient of x4

coefTestH(linmod,c("x1=x2","x3=2*x4"))

The F-statistic and p-value for the hypothesis test are reported.
}

