% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_hypothesis_functions.R
\name{coefTestHAC}
\alias{coefTestHAC}
\title{Linear Hypothesis Test (Time Series Data)}
\usage{
coefTestHAC(model, conditions)
}
\arguments{
\item{model}{Regression model being tested.}

\item{conditions}{Hypothesis being tested.}
}
\description{
Tests a linear hypothesis about coefficients. The calculation corrects for possible presence of heteroskedasticity and autocorrelation using the method developed by Newey and West. This method is appropriate for time series data.
}
\section{Details}{

Let the regression model be: linmod=lm(y~x1+x2+x3+x4)

To test the hypothesis that the coefficient of x1 equals the coefficient of x2 and that the coefficient of x3 equals twice the coefficient of x4

coefTestHAC(linmod,c("x1=x2","x3=2*x4"))

The F-statistic and p-value for the hypothesis test are reported.
}

