% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlChart_functions.R
\name{controlChart}
\alias{controlChart}
\title{Control Chart}
\usage{
controlChart(chartdata, mu, sig, n, alpha)
}
\arguments{
\item{chartdata}{The name of the variable to be plotted.}

\item{mu}{The anticipated mean when the production process is working properly.}

\item{sig}{The anticipated standard deviation when the process is working properly.}

\item{n}{The number of observations per sample.}

\item{alpha}{The significance level.}
}
\description{
This function plots a control chart for monitoring a production process for a product with a continuous measure (e.g., weight of a package).
}
