% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlChart_functions.R
\name{controlChartBinary}
\alias{controlChartBinary}
\title{Control Chart Binary}
\usage{
controlChartBinary(chartdata, p, n, alpha)
}
\arguments{
\item{chartdata}{The name of the variable to be plotted.}

\item{p}{The target probability for the outcome expressed as a value between 0 and 1.}

\item{n}{The number of observations per sample.}

\item{alpha}{The significance level.}
}
\description{
This function plots a control chart for monitoring a production process for binary outcomes.
}
