% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticities.R
\name{elasticities}
\alias{elasticities}
\title{Coefficient Elasticities}
\usage{
elasticities(linmod)
}
\arguments{
\item{linmod}{The linear regression model.}
}
\description{
Elasticities evaluated at the means of the variables are calculated for the coefficients of a linear regression model.
}
