% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fit_distribution}
\alias{fit_distribution}
\title{Fit Distribution}
\usage{
fit_distribution(.data)
}
\arguments{
\item{.data}{Data.}
}
\description{
Fits up to 8 different distrubtions to a set of data and recommends the best-fitting distribution, where best-fitting is considered to be the distribution with the smallest AIC value. Prints the distribution parameters and R instructions for sampling the recommended distribution. Outputs a histogram of the data overlayed by the density function of the recommended distribution.
}
