% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fitdist_exponential}
\alias{fitdist_exponential}
\title{Fit Distribution Exponential}
\usage{
fitdist_exponential(my_data)
}
\arguments{
\item{my_data}{The data.}
}
\value{
List of the fitted distribution parameters.
}
\description{
Fits the exponentional distribution to the given data. Adjusts the scaling for estimation and returns the estimate parameters at the original scaling.
}
