% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fitdist_parameters_skew_normal}
\alias{fitdist_parameters_skew_normal}
\title{Fit Distribution Parameters Skew Normal}
\usage{
fitdist_parameters_skew_normal(my_data)
}
\arguments{
\item{my_data}{The data.}
}
\value{
List of distribution parameters.
}
\description{
Fits the skew normal distribution to the data.
}
