% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fitdist_parameters_t}
\alias{fitdist_parameters_t}
\title{Fit Distribution Parameters t}
\usage{
fitdist_parameters_t(my_data)
}
\arguments{
\item{my_data}{The data.}
}
\value{
List of the distribution parameters.
}
\description{
Fits the t distribution to the given data.
}
