% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fitdist_uniform}
\alias{fitdist_uniform}
\title{Fit Distribution Uniform}
\usage{
fitdist_uniform(.data)
}
\arguments{
\item{.data}{The data.}
}
\value{
List of parameters from the uniform distribution.
}
\description{
Fits the uniform distribution to the given data.
}
