% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_functions.R
\name{go3DPlot}
\alias{go3DPlot}
\title{3D Plot}
\usage{
go3DPlot(varx, vary, varz, theTheta, thePhi, data)
}
\arguments{
\item{varx}{name of variable on one horizontal axis. Use quotes.}

\item{vary}{name of variable on other horizontal axis. Use quotes.}

\item{varz}{name of variable on vertical axis. Use quotes.}

\item{theTheta}{Theta value for viewing angle.}

\item{thePhi}{Phi value for viewing angle.}

\item{data}{The dataframe containing data.}
}
\description{
3D scatter plot with regression plane.
}
\section{Details}{

Each of the commands below produces a plot of the data and regression plane with x1 and x2 on the horizontal axes and y on the vertical axis. The first command produces a plot with viewing angle theta=50 and phi=15. The second produces a plot with viewing angle theta=60 and phi=20.

go3DPlot("x1","x2","y",50,15,data=Plot_3D_Example)

go3DPlot("x1","x2","y",60,20,data=Plot_3D_Example)
}

