% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_estimation.R
\name{goLASSO}
\alias{goLASSO}
\title{LASSO}
\usage{
goLASSO(datafile, predDat)
}
\arguments{
\item{datafile}{Used for estimating model.}

\item{predDat}{Used for prediction.}
}
\value{
List of results.
}
\description{
Performs LASSO regression analysis.
}
