% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_CI.R
\name{hist_CI}
\alias{hist_CI}
\title{Histogram with Confidence Interval}
\usage{
hist_CI(.data, alpha, title, xlabel, breaks = 20)
}
\arguments{
\item{.data}{The data.}

\item{alpha}{Alpha}

\item{title}{Graph title.}

\item{xlabel}{Graph x label.}

\item{breaks}{Histogram breaks. Defaults to 20.}
}
\description{
Plots a histogram with confidence interval.
}
