% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{make_histogram}
\alias{make_histogram}
\title{Make Histogram}
\usage{
make_histogram(.data, dist)
}
\arguments{
\item{.data}{The data.}

\item{dist}{List of distribution parameters.}
}
\description{
Outputs a histogram of the data overlayed by the density function of the recommended distribution.
}
