% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actual_fitted.R
\name{plotActualFitted}
\alias{plotActualFitted}
\title{Plot Actual Fitted}
\usage{
plotActualFitted(linmod, MyData, col_name)
}
\arguments{
\item{linmod}{Linear model}

\item{MyData}{The data}

\item{col_name}{Name of column variable as a string (use quotes).}
}
\description{
Plots actual and fitted values of the dependent variable in a regression and 95% prediction interval. The horizontal axis is a sequence of integers denoting observations.
}
