% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_actual_fitted.R
\name{plotLogActualFitted}
\alias{plotLogActualFitted}
\title{Plot Log Actual Fitted}
\usage{
plotLogActualFitted(linmod, MyData, col_name)
}
\arguments{
\item{linmod}{Linear model.}

\item{MyData}{The data.}

\item{col_name}{Name of column variable as a string (use quotes).}
}
\description{
Plot Log Actual Fitted
}
