% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_objective_function.R
\name{plotObjectiveFunction}
\alias{plotObjectiveFunction}
\title{Plot Objective Function}
\usage{
plotObjectiveFunction(fn, from, to)
}
\arguments{
\item{fn}{A function.}

\item{from}{Minimum x value.}

\item{to}{Maximum x value.}
}
\value{
results A table.
}
\description{
Plots x to mean of the input function over the input range.
}
