% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{print_distribution_instructions}
\alias{print_distribution_instructions}
\title{Print Distribution Instructions}
\usage{
print_distribution_instructions(dist)
}
\arguments{
\item{dist}{List of parameters from the fitted distribution.}
}
\description{
Prints R instructions for sampling the recommended distribution.
}
