% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{print_distribution_parameters}
\alias{print_distribution_parameters}
\title{Print Distribution Parameters}
\usage{
print_distribution_parameters(dist)
}
\arguments{
\item{dist}{List of parameters from the fitted distribution.}
}
\description{
Prints the name of the recommended distribution and the distribution parameters.
}
