% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramsey_test.R
\name{ramseyTest}
\alias{ramseyTest}
\title{Ramsey Specification Test}
\usage{
ramseyTest(linmod)
}
\arguments{
\item{linmod}{Linear model.}
}
\description{
This function conducts the Ramsey Regression Specification Error Test (RESET) for a regression model. The function conducts the Ramsey test with one fitted term and then with two fitted terms. The p-values for both tests are reported.
}
