% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_coefficients.R
\name{scaledCoefficients}
\alias{scaledCoefficients}
\title{Scaled Coefficients}
\usage{
scaledCoefficients(linmod)
}
\arguments{
\item{linmod}{The linear regression equation for which scaled coefficients are to be calculated.}
}
\description{
Calculates standardized coefficients (also known as beta weights) and elasticities. (A standardized coefficient is the coefficient obtained from a regression in which both the independent variable and the dependent variable are standardized to have mean equal to zero and standard deviation equal to one.) The function reports three columns which contain the coefficients, the standardized coefficients, and the elasticities evaluated at the means of the variables.
}
