% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{tprstats setup}
\usage{
setup(suppress = TRUE)
}
\arguments{
\item{suppress}{Suppresses warnings for students by default. Suppress can be set to false for troubleshooting.}
}
\description{
Installs course packages not already installed as tprstats dependencies.
}
