% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdcoefs.R
\name{stdcoefs}
\alias{stdcoefs}
\title{Standardized Regression Coefficients}
\usage{
stdcoefs(linmod)
}
\arguments{
\item{linmod}{The linear regression model.}
}
\description{
Calculates standardized regression coefficients, also known as beta weights. A standardized coefficient is the coefficient obtained from a regression in which both the independent variable and the dependent variable are standardized to have mean equal to zero and standard deviation equal to one.
}
