% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summaryArima}
\alias{summaryArima}
\title{Summary for Regression with Autoregressive Errors}
\usage{
summaryArima(arima_mod)
}
\arguments{
\item{arima_mod}{The regression model with autoregressive errors.}
}
\description{
Displays results of regression estimated with autoregressive errors. The function will also provide a summary for a regression with both autoregressive and moving average terms.
}
\section{Details}{

Consider a regression of y on x1, x2, and x3 that is estimated with second-order autoregressive errors as follows: \cr
ar2Reg=with(mydata, Arima(y, xreg=cbind(x1,x2,x3), order=c(2,0,0), include.constant = TRUE))

The results of this regression are then reported by the following command: \cr
summaryArima(ar2Reg)

Consider a regression of y on x1, x2, and x3 that is estimated with both first-order autoregressive and first-order moving average errors as follows: \cr
armaReg=with(mydata, Arima(y, xreg=cbind(x1,x2,x3), order=c(1,0,1), include.constant = TRUE))

The results of this regression are then reported by the following command: \cr
summaryArima(armaReg)
}

