% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summaryH}
\alias{summaryH}
\title{Summary of Regression (Cross Section Data)}
\usage{
summaryH(model)
}
\arguments{
\item{model}{The linear regression model.}
}
\value{
sumry Returns a summary.lm object with Heteroscedasticity-consistent standard errors
}
\description{
This function displays regression results with standard errors adjusted for possible presence of heteroskedasticity using the method developed by White. This method is appropriate for cross-section data.
}
