% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summaryHAC}
\alias{summaryHAC}
\title{Summary of Regression (Time Series Data)}
\usage{
summaryHAC(model)
}
\arguments{
\item{model}{The linear regression model.}
}
\description{
This function displays regression results with standard errors adjusted for possible presence of heteroskedasticity and autocorrelation using the method developed by Newey and West.  This method is appropriate for time series data.
}
