% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summaryStats}
\alias{summaryStats}
\title{Summary Statistics}
\usage{
summaryStats(.data)
}
\arguments{
\item{.data}{The data.}
}
\description{
Returns mean, median, max, min and std. deviation for each variable in the data.
}
