% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{test_select_distribution}
\alias{test_select_distribution}
\title{Test Best-fitting Distribution}
\usage{
test_select_distribution()
}
\description{
Tests fit distribution against a variety of generated data.
}
