% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distribution.R
\name{fit_distribution}
\alias{fit_distribution}
\title{Fit Distribution}
\usage{
fit_distribution(
  .data,
  distname = c("norm", "unif", "t", "snorm", "weibull", "gamma", "lnorm", "exp")
)
}
\arguments{
\item{.data}{The data.}

\item{distname}{The distribution name.}
}
\description{
Fits a distribution to data.The abbreviations for distribution names are: "norm", "unif", "t", "snorm", "weibull", "gamma", "lnorm", "exp"
}
