% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_CI.R
\name{hist_CI}
\alias{hist_CI}
\title{Histogram with Confidence Interval}
\usage{
hist_CI(
  .data,
  alpha = 0.1,
  main = "Histogram",
  xlab = "Data and (1-alpha)\% Confidence Interval",
  breaks = 20
)
}
\arguments{
\item{.data}{The data.}

\item{alpha}{Alpha value. Defaults to 0.1.}

\item{main}{Graph title.}

\item{xlab}{Graph x label.}

\item{breaks}{Histogram breaks. Defaults to 20.}
}
\description{
Plots a histogram with confidence interval.
}
