% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{tprstats setup}
\usage{
setup(load = TRUE)
}
\arguments{
\item{load}{Determines whether packages should be loaded. Defaults to TRUE.}
}
\description{
Installs additional course packages not already installed as tprstats dependencies. Loads these packages if load = TRUE.
}
