% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_tprstats}
\alias{setup_tprstats}
\title{Deprecated Setup for tprstats}
\usage{
setup_tprstats(load = TRUE)
}
\arguments{
\item{load}{Indicates of package should be loaded (load = TRUE) or if setup should just check the packages have been installed (load = FALSE)}
}
\description{
This command is now identical to the tprstats::setup() command. This command just calls tprstats::setup().
}
