% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_tprstats}
\alias{setup_tprstats}
\title{Deprecated Setup for tprstats}
\usage{
setup_tprstats()
}
\description{
This command is now identical to the tprstats::setup() command. This command just calls tprstats::setup().
}
